db/requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
pgvector==0.2.4
numpy>=1.21.0 